+++ Projectiles/WeaponProjectiles/LeafSpawner.cs
@@ -4,7 +4,7 @@ using Terraria.ModLoader;
 
 namespace StarlightRiver.Projectiles.WeaponProjectiles
 {
-    internal class LeafSpawner : ModProjectile
+    class LeafSpawner : ModProjectile
     {
         public override void SetDefaults()
         {
@@ -15,20 +15,13 @@ namespace StarlightRiver.Projectiles.WeaponProjectiles
             projectile.ignoreWater = true;
             projectile.friendly = true;
         }
-        public int Proj { get; set; }
+        public int proj { get; set; }
         public override void AI()
         {
             projectile.ai[0]++;
-            if (!Main.projectile[Proj].active)
-            {
-                projectile.Kill();
-            }
-
-            projectile.position = Main.projectile[Proj].position;
-            if (projectile.ai[0] % 10 == 0)
-            {
-                Projectile.NewProjectile(projectile.Center, new Vector2(0, 0), ModContent.ProjectileType<Leaf>(), projectile.damage, projectile.knockBack, projectile.owner);
-            }
+            if (!Main.projectile[proj].active) projectile.Kill();
+            projectile.position = Main.projectile[proj].position;
+            if (projectile.ai[0] % 10 == 0) Projectile.NewProjectile(projectile.Center, new Vector2(0, 0), ModContent.ProjectileType<Leaf>(), projectile.damage, projectile.knockBack, projectile.owner);
         }
     }
 }
